requirements.txt
playwright
beautifulsoup4
httpx
python-dotenv
pyotp
